Add Linux specific versions to tasks.json
All vscode tasks now have a Linux version, with the exception of:

- Set input path
- Publish slides to Github Pages
- Paper to DOCX specific commit
- Paper track changes DOCX

The latter depends on Office Word being installed on the system.
It is impossible to port that to Linux.
The other tasks should have a linux version.

There might be a possibilty to create a track changes document with LibreOffice.